refactor: Extract MDX types to an external file
