+++ design-system/components/project/project.types.tsx
@@ -1,5 +1,6 @@
-import { Dictionary, Project } from '@/scripts'
+import { Dictionary } from '@/scripts'
 import { Locale } from '@/i18n.config'
+import type { Project } from '@/types'
 
 export type ProjectProps = Pick<
   Project,
