+++ scripts/index.ts
@@ -1,9 +1,5 @@
 export { type Dictionary, getDictionary } from './dictionaries'
 export {
-  type Page,
-  type Pagination,
-  type Post,
-  type Project,
   createMDXPagination,
   getMDX,
   getMDXes,
