+++ types/index.tsx
@@ -1,5 +1,14 @@
 export { type MetaFontOptions } from './meta-font'
 export {
+  type Frontmatter,
+  type MDX,
+  type MDXTypes,
+  type Page,
+  type Pagination,
+  type Post,
+  type Project
+} from './mdx'
+export {
   type PageProps,
   type NestedPageProps,
   type SearchPageProps
